DemoApp/src/app/custom/home/home/home.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { HomeComponent } from './home.component';
import { ReversePipe } from '../../pipe/reverse.pipe';

describe('HomeComponent', () => {
  let component: HomeComponent;
  let fixture: ComponentFixture<HomeComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [HomeComponent, ReversePipe]
    })
    .compileComponents();

    fixture = TestBed.createComponent(HomeComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  fit("should call console log", ()=>{
    spyOn(console, 'log');
    component.onSubmit(component.userForm);
    expect(console.log).toHaveBeenCalled
    expect(22).toBe(22)
  })
});
